+++ include/libpldm/bios_table.h
@@ -655,8 +655,19 @@ size_t pldm_bios_table_pad_checksum_size(size_t size_without_pad);
  *  @param[in] size_without_pad - Table size without pad and checksum
  *  @return Total size of the table
  */
-size_t pldm_bios_table_append_pad_checksum(void *table, size_t size,
-					   size_t size_without_pad);
+size_t pldm_bios_table_append_pad_checksum(void *table, size_t capacity,
+					   size_t size);
+
+/** @brief Append pad and checksum at the end of the table or return an error
+ *  @param[in,out] table - Pointer to a buffer of a bios table
+ *  @param[in] capacity - Size of the buffer of a bios table
+ *  @param[in,out] size - On input, the table size without pad and checksum, on output, the table
+ *  			  with the padding and checksum appended
+ *  @return Total size of the table on success, otherwise PLDM_ERROR_INVALID_LENGTH if size lacks
+ *          capacity to encode the checksum in the buffer provided by table
+ */
+int pldm_bios_table_append_pad_checksum_check(void *table, size_t capacity,
+					      size_t *size);
 
 /** @brief Build a new table and update an entry
  *  @param[in] src_table - Pointer to the source table