+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Change categories:
 
 ## [Unreleased]
 
+### Added
+
+1. bios_table: Introduce pldm_bios_table_append_pad_checksum_check()
+
 ### Changed
 
 1. bios_table: pldm_bios_table_string_entry_encode_check(): Handle overflow